Abstract StudyInfo and StudySize out of StudyPage
